requirements.txt
# API service
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
huggingface_hub>=0.25.0

# Deep Learning
tensorflow-cpu>=2.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
